Rename create_resume.py to tailor_resume.py
